+++ src/common/ui/LogoTitle/LogoTitle.tsx
@@ -1,10 +1,15 @@
 import React from "react";
 import { ImgLogo, AdminSite } from "src/common/ui/MainDivContainer";
 
-export const LogoTitle: React.SFC = (props)=>{
-    return (
-    <div>
-        <ImgLogo src={require("src/images/poc-chef-logo.svg")} />
-        <AdminSite>ADMIN SITE</AdminSite>
-    </div>);
+export interface LogoTitleProps {
+  title: boolean;
+  titleText?: string;
 }
+export const LogoTitle: React.SFC<LogoTitleProps> = ({ title, titleText }) => {
+  return (
+    <div style={{ textAlign: "center" }}>
+      <ImgLogo src={require("src/images/poc-chef-logo.svg")} />
+      {title && <AdminSite>{titleText}</AdminSite>}
+    </div>
+  );
+};
