+++ src/common/views/Splash/Splash.tsx
@@ -3,13 +3,15 @@ import { MainDivContainer, ImgLogo } from "../../ui/MainDivContainer";
 import { RouteComponentProps } from "react-router";
 import { loginPartnerRoute } from "src/partner/routes";
 import { loginUserRoute } from "src/user/routes";
+import { LogoTitle } from "src/common/ui/LogoTitle";
 
 export const Splash: React.FC<RouteComponentProps> = ({ history, location }) => {
+  const { pathname } = location;
+  const isPartnerRoute = /^\/partner/gi.test(pathname);
+  const isUserRoute = /^\/user/gi.test(pathname);
+
   useEffect(() => {
     setTimeout(() => {
-      const { pathname } = location;
-      const isPartnerRoute = /^\/partner/gi.test(pathname);
-      const isUserRoute = /^\/user/gi.test(pathname);
       const nextLocation = isPartnerRoute
         ? loginPartnerRoute
         : isUserRoute
@@ -20,9 +22,14 @@ export const Splash: React.FC<RouteComponentProps> = ({ history, location }) => 
     }, 2000);
   }, []);
 
+  const MainDivContainerProps = isPartnerRoute ? { partner: true } : { user: true };
+  const LogoTitleProps = isPartnerRoute
+    ? { title: true, titleText: "ADMIN SITE" }
+    : { title: false };
+
   return (
-    <MainDivContainer>
-      <ImgLogo src={require("src/images/poc-chef-logo.svg")} />
+    <MainDivContainer {...MainDivContainerProps}>
+      <LogoTitle {...LogoTitleProps} />
     </MainDivContainer>
   );
 };
